test: fuse the lands and land-lease specs into one file
both files were testing the lands component now that land-lease was
folded in last commit, and both used the same createLandsComponent
factory with overlapping mock setup. merging them removes a layer of
duplication, lets every test share one beforeEach, and gives the
component a single canonical spec.

the new test/unit/lands.spec.ts has one outer describe and sub-describes
per method group: getLandPermissions, getLandOperators, cachedFetch
reuse with the real cachedFetch component, hasLandLease,
getAuthorizations caching and dedup, refreshAuthorizations, and
getLeaseHoldersForParcels.

setup keeps the two transports separate by name — mockLambdasFetch
for the lambdas (cachedFetch) calls and mockLeaseFetch for the lease
authorizations document — so each test only configures the transport it
exercises.

deletes:
- test/integration/land-component.spec.ts (also moved out of the
  integration tree; it was always unit-style with mocked deps)
- test/unit/lands-lease.spec.ts
