+++ test/unit/lands.spec.ts
@@ -1,0 +1,531 @@
+import {
+  createLandsComponent,
+  ILandComponent,
+  LandsParcelOperatorsResponse,
+  LandsParcelPermissionsResponse
+} from '../../src/adapters/lands'
+import { cachedFetchComponent } from '../../src/adapters/fetch'
+import { createLoggerMockedComponent } from '../mocks/logger-mock'
+
+// The lands component fans out to two different transports:
+//   - cachedFetch (for parcel permissions / operators against the lambdas API)
+//   - fetch       (for the off-chain lease authorizations document)
+// Tests configure them independently via the mocks below.
+
+const SAMPLE_AUTHORIZATIONS = [
+  {
+    name: 'Test Authorization 1',
+    desc: 'Test Description 1',
+    contactInfo: { name: 'Test Contact 1' },
+    addresses: ['0x1234567890123456789012345678901234567890'],
+    plots: ['-73,50', '10,20']
+  },
+  {
+    name: 'Test Authorization 2',
+    desc: 'Test Description 2',
+    contactInfo: { name: 'Test Contact 2' },
+    addresses: ['0xabcdefabcdefabcdefabcdefabcdefabcdefabcd'],
+    plots: ['-100,100']
+  }
+]
+
+describe('lands component', () => {
+  let lands: ILandComponent
+  let mockLambdasFetch: jest.Mock
+  let mockLeaseFetch: { fetch: jest.Mock }
+
+  beforeEach(async () => {
+    jest.clearAllMocks()
+
+    mockLambdasFetch = jest.fn()
+    mockLeaseFetch = { fetch: jest.fn() }
+
+    lands = await createLandsComponent({
+      config: {
+        requireString: jest
+          .fn()
+          .mockImplementation((key: string) =>
+            Promise.resolve(key === 'LAMBDAS_URL' ? 'https://lambdas.decentraland.org/api' : '')
+          )
+      } as any,
+      cachedFetch: {
+        cache: jest.fn().mockReturnValue({ fetch: mockLambdasFetch })
+      } as any,
+      fetch: mockLeaseFetch as any,
+      logs: createLoggerMockedComponent()
+    })
+  })
+
+  describe('getLandPermissions', () => {
+    it('should return owner=true when user has owner permission', async () => {
+      const response: LandsParcelPermissionsResponse = {
+        owner: true,
+        operator: false,
+        updateOperator: false,
+        updateManager: false,
+        approvedForAll: false
+      }
+      mockLambdasFetch.mockResolvedValueOnce(response)
+
+      const result = await lands.getLandPermissions('0xUserAddress', ['10,20', '99,99'])
+
+      expect(result).toEqual(response)
+      expect(mockLambdasFetch).toHaveBeenCalledWith(
+        'https://lambdas.decentraland.org/api/users/0xuseraddress/parcels/10/20/permissions'
+      )
+    })
+
+    it('should return operator=true when user has operator permission', async () => {
+      const response: LandsParcelPermissionsResponse = {
+        owner: false,
+        operator: true,
+        updateOperator: false,
+        updateManager: false,
+        approvedForAll: false
+      }
+      mockLambdasFetch.mockResolvedValueOnce(response)
+
+      const result = await lands.getLandPermissions('0xUserAddress', ['50,60', '99,99'])
+
+      expect(result).toEqual(response)
+      expect(mockLambdasFetch).toHaveBeenCalledWith(
+        'https://lambdas.decentraland.org/api/users/0xuseraddress/parcels/50/60/permissions'
+      )
+    })
+
+    it('should throw LandPermissionsNotFoundError when the response is empty', async () => {
+      mockLambdasFetch.mockResolvedValueOnce(null)
+
+      await expect(lands.getLandPermissions('0xUserAddress', ['10,20'])).rejects.toThrow(
+        'Land permissions not found for 0xUserAddress at 10,20'
+      )
+    })
+
+    it('should throw when LAMBDAS_URL is not configured', async () => {
+      const componentWithoutUrl = await createLandsComponent({
+        config: { requireString: jest.fn().mockResolvedValue('') } as any,
+        cachedFetch: { cache: jest.fn().mockReturnValue({ fetch: jest.fn() }) } as any,
+        fetch: { fetch: jest.fn() } as any,
+        logs: createLoggerMockedComponent()
+      })
+
+      await expect(componentWithoutUrl.getLandPermissions('0xUserAddress', ['10,20'])).rejects.toThrow(
+        'Lambdas URL is not set'
+      )
+    })
+  })
+
+  describe('getLandOperators', () => {
+    it('should return owner and operator when both exist', async () => {
+      const response: LandsParcelOperatorsResponse = {
+        owner: '0xOwnerAddress',
+        operator: '0xOperatorAddress',
+        updateOperator: null,
+        updateManagers: [],
+        approvedForAll: []
+      }
+      mockLambdasFetch.mockResolvedValueOnce(response)
+
+      const result = await lands.getLandOperators('10,20')
+
+      expect(result).toEqual(response)
+      expect(mockLambdasFetch).toHaveBeenCalledWith('https://lambdas.decentraland.org/api/parcels/10/20/operators')
+    })
+
+    it('should return updateOperator when it exists', async () => {
+      mockLambdasFetch.mockResolvedValueOnce({
+        owner: '0xOwnerAddress',
+        operator: null,
+        updateOperator: '0xUpdateOperatorAddress',
+        updateManagers: [],
+        approvedForAll: []
+      })
+
+      const result = await lands.getLandOperators('10,20')
+
+      expect(result.updateOperator).toBe('0xUpdateOperatorAddress')
+    })
+
+    it('should return updateManagers when present', async () => {
+      mockLambdasFetch.mockResolvedValueOnce({
+        owner: '0xOwnerAddress',
+        operator: null,
+        updateOperator: null,
+        updateManagers: ['0xUpdateManagerAddress'],
+        approvedForAll: []
+      })
+
+      const result = await lands.getLandOperators('10,20')
+
+      expect(result.updateManagers).toEqual(['0xUpdateManagerAddress'])
+    })
+
+    it('should return approvedForAll when present', async () => {
+      mockLambdasFetch.mockResolvedValueOnce({
+        owner: '0xOwnerAddress',
+        operator: null,
+        updateOperator: null,
+        updateManagers: [],
+        approvedForAll: ['0xApprovedForAllAddress']
+      })
+
+      const result = await lands.getLandOperators('10,20')
+
+      expect(result.approvedForAll).toEqual(['0xApprovedForAllAddress'])
+    })
+
+    it('should throw LandPermissionsNotFoundError when the response is empty', async () => {
+      mockLambdasFetch.mockResolvedValueOnce(null)
+
+      await expect(lands.getLandOperators('10,20')).rejects.toThrow('Land permissions not found for 10,20')
+    })
+
+    it('should throw when LAMBDAS_URL is not configured', async () => {
+      const componentWithoutUrl = await createLandsComponent({
+        config: { requireString: jest.fn().mockResolvedValue('') } as any,
+        cachedFetch: { cache: jest.fn().mockReturnValue({ fetch: jest.fn() }) } as any,
+        fetch: { fetch: jest.fn() } as any,
+        logs: createLoggerMockedComponent()
+      })
+
+      await expect(componentWithoutUrl.getLandOperators('10,20')).rejects.toThrow('Lambdas URL is not set')
+    })
+  })
+
+  describe('cachedFetch reuse with the real cachedFetch component', () => {
+    let cachedFetchFetch: jest.Mock
+    let cachedLands: ILandComponent
+
+    beforeEach(async () => {
+      cachedFetchFetch = jest.fn()
+      const realCachedFetch = await cachedFetchComponent({
+        fetch: { fetch: cachedFetchFetch },
+        logs: createLoggerMockedComponent()
+      })
+
+      cachedLands = await createLandsComponent({
+        config: {
+          requireString: jest.fn().mockResolvedValue('https://lambdas.decentraland.org/api')
+        } as any,
+        cachedFetch: realCachedFetch,
+        fetch: { fetch: jest.fn() } as any,
+        logs: createLoggerMockedComponent()
+      })
+    })
+
+    it('should call the upstream only once for two getLandPermissions calls with the same key', async () => {
+      cachedFetchFetch.mockResolvedValueOnce({
+        ok: true,
+        json: jest.fn().mockResolvedValue({
+          owner: true,
+          operator: false,
+          updateOperator: false,
+          updateManager: false,
+          approvedForAll: false
+        })
+      })
+
+      await cachedLands.getLandPermissions('0xUserAddress', ['10,20'])
+      await cachedLands.getLandPermissions('0xUserAddress', ['10,20'])
+
+      expect(cachedFetchFetch).toHaveBeenCalledTimes(1)
+    })
+
+    it('should share the cache entry across address casings', async () => {
+      cachedFetchFetch.mockResolvedValueOnce({
+        ok: true,
+        json: jest.fn().mockResolvedValue({
+          owner: true,
+          operator: false,
+          updateOperator: false,
+          updateManager: false,
+          approvedForAll: false
+        })
+      })
+
+      await cachedLands.getLandPermissions('0xUserAddress', ['10,20'])
+      await cachedLands.getLandPermissions('0xuseraddress', ['10,20'])
+
+      expect(cachedFetchFetch).toHaveBeenCalledTimes(1)
+    })
+
+    it('should call the upstream only once for two getLandOperators calls with the same parcel', async () => {
+      cachedFetchFetch.mockResolvedValueOnce({
+        ok: true,
+        json: jest.fn().mockResolvedValue({
+          owner: '0xOwnerAddress',
+          operator: null,
+          updateOperator: null,
+          updateManagers: [],
+          approvedForAll: []
+        })
+      })
+
+      await cachedLands.getLandOperators('10,20')
+      await cachedLands.getLandOperators('10,20')
+
+      expect(cachedFetchFetch).toHaveBeenCalledTimes(1)
+    })
+  })
+
+  describe('hasLandLease', () => {
+    describe('and the user has a lease for the parcel', () => {
+      beforeEach(() => {
+        mockLeaseFetch.fetch.mockResolvedValue({
+          ok: true,
+          json: () => Promise.resolve(SAMPLE_AUTHORIZATIONS)
+        })
+      })
+
+      it('should return true', async () => {
+        const result = await lands.hasLandLease('0x1234567890123456789012345678901234567890', ['-73,50'])
+        expect(result).toBe(true)
+      })
+
+      it('should return true with case-insensitive address matching', async () => {
+        const result = await lands.hasLandLease('0X1234567890123456789012345678901234567890', ['-73,50'])
+        expect(result).toBe(true)
+      })
+    })
+
+    describe('and the user has no overlap with their authorized parcels', () => {
+      beforeEach(() => {
+        mockLeaseFetch.fetch.mockResolvedValue({
+          ok: true,
+          json: () => Promise.resolve(SAMPLE_AUTHORIZATIONS)
+        })
+      })
+
+      it('should return false', async () => {
+        const result = await lands.hasLandLease('0x1234567890123456789012345678901234567890', ['999,999'])
+        expect(result).toBe(false)
+      })
+    })
+
+    describe('and the user is not in any authorization', () => {
+      beforeEach(() => {
+        mockLeaseFetch.fetch.mockResolvedValue({
+          ok: true,
+          json: () => Promise.resolve(SAMPLE_AUTHORIZATIONS)
+        })
+      })
+
+      it('should return false', async () => {
+        const result = await lands.hasLandLease('0x9999999999999999999999999999999999999999', ['-73,50'])
+        expect(result).toBe(false)
+      })
+    })
+
+    describe('and the parcels list is empty', () => {
+      beforeEach(() => {
+        mockLeaseFetch.fetch.mockResolvedValue({
+          ok: true,
+          json: () => Promise.resolve(SAMPLE_AUTHORIZATIONS)
+        })
+      })
+
+      it('should return false', async () => {
+        const result = await lands.hasLandLease('0x1234567890123456789012345678901234567890', [])
+        expect(result).toBe(false)
+      })
+    })
+
+    describe('and the upstream fetch fails', () => {
+      beforeEach(() => {
+        mockLeaseFetch.fetch.mockRejectedValue(new Error('Network error'))
+      })
+
+      it('should return false', async () => {
+        const result = await lands.hasLandLease('0x1234567890123456789012345678901234567890', ['-73,50'])
+        expect(result).toBe(false)
+      })
+    })
+
+    describe('and the upstream returns a non-ok response', () => {
+      beforeEach(() => {
+        mockLeaseFetch.fetch.mockResolvedValue({ ok: false, status: 404, statusText: 'Not Found' })
+      })
+
+      it('should return false', async () => {
+        const result = await lands.hasLandLease('0x1234567890123456789012345678901234567890', ['-73,50'])
+        expect(result).toBe(false)
+      })
+    })
+  })
+
+  describe('getAuthorizations caching and dedup', () => {
+    describe('and the cache is valid', () => {
+      beforeEach(() => {
+        mockLeaseFetch.fetch.mockResolvedValue({
+          ok: true,
+          json: () => Promise.resolve(SAMPLE_AUTHORIZATIONS)
+        })
+      })
+
+      it('should hit the upstream only once across multiple calls within the TTL', async () => {
+        await lands.hasLandLease('0x1234567890123456789012345678901234567890', ['-73,50'])
+        await lands.hasLandLease('0xabcdefabcdefabcdefabcdefabcdefabcdefabcd', ['-100,100'])
+        expect(mockLeaseFetch.fetch).toHaveBeenCalledTimes(1)
+      })
+    })
+
+    describe('and the cache has expired', () => {
+      beforeEach(() => {
+        jest.useFakeTimers()
+        mockLeaseFetch.fetch.mockResolvedValue({
+          ok: true,
+          json: () => Promise.resolve(SAMPLE_AUTHORIZATIONS)
+        })
+      })
+
+      afterEach(() => {
+        jest.useRealTimers()
+      })
+
+      it('should refetch fresh data after TTL elapses', async () => {
+        await lands.hasLandLease('0x1234567890123456789012345678901234567890', ['-73,50'])
+        jest.advanceTimersByTime(5 * 60 * 1000 + 1000)
+        await lands.hasLandLease('0xabcdefabcdefabcdefabcdefabcdefabcdefabcd', ['-100,100'])
+        expect(mockLeaseFetch.fetch).toHaveBeenCalledTimes(2)
+      })
+    })
+
+    describe('and two callers request authorizations concurrently', () => {
+      let firstResult: any
+      let secondResult: any
+
+      beforeEach(async () => {
+        mockLeaseFetch.fetch.mockImplementationOnce(
+          () =>
+            new Promise((resolve) =>
+              setTimeout(() => resolve({ ok: true, json: () => Promise.resolve(SAMPLE_AUTHORIZATIONS) }), 0)
+            )
+        )
+        ;[firstResult, secondResult] = await Promise.all([lands.getAuthorizations(), lands.getAuthorizations()])
+      })
+
+      it('should issue only one upstream fetch (in-flight dedup)', () => {
+        expect(mockLeaseFetch.fetch).toHaveBeenCalledTimes(1)
+      })
+
+      it('should resolve both callers with the same payload', () => {
+        expect(firstResult).toEqual(secondResult)
+        expect(firstResult.authorizations).toHaveLength(SAMPLE_AUTHORIZATIONS.length)
+      })
+    })
+
+    describe('and a refresh fetch fails after the cache has expired', () => {
+      let result: any
+
+      beforeEach(async () => {
+        jest.useFakeTimers()
+        mockLeaseFetch.fetch.mockResolvedValueOnce({
+          ok: true,
+          json: () => Promise.resolve(SAMPLE_AUTHORIZATIONS)
+        })
+
+        // Populate cache
+        await lands.getAuthorizations()
+        // Expire it
+        jest.advanceTimersByTime(5 * 60 * 1000 + 1000)
+        // Next fetch fails
+        mockLeaseFetch.fetch.mockRejectedValueOnce(new Error('Network down'))
+
+        result = await lands.getAuthorizations()
+      })
+
+      afterEach(() => {
+        jest.useRealTimers()
+      })
+
+      it('should serve stale cache instead of an empty result', () => {
+        expect(result.authorizations).toHaveLength(SAMPLE_AUTHORIZATIONS.length)
+      })
+
+      it('should have attempted the fresh fetch', () => {
+        expect(mockLeaseFetch.fetch).toHaveBeenCalledTimes(2)
+      })
+    })
+
+    describe('and the very first fetch fails with no cache populated', () => {
+      let result: any
+
+      beforeEach(async () => {
+        mockLeaseFetch.fetch.mockRejectedValueOnce(new Error('Network down'))
+        result = await lands.getAuthorizations()
+      })
+
+      it('should return an empty authorizations payload', () => {
+        expect(result).toEqual({ authorizations: [] })
+      })
+    })
+  })
+
+  describe('refreshAuthorizations', () => {
+    beforeEach(() => {
+      mockLeaseFetch.fetch.mockResolvedValue({
+        ok: true,
+        json: () => Promise.resolve(SAMPLE_AUTHORIZATIONS)
+      })
+    })
+
+    it('should clear the cache and trigger a fresh fetch', async () => {
+      await lands.hasLandLease('0x1234567890123456789012345678901234567890', ['-73,50'])
+      await lands.refreshAuthorizations()
+      await lands.hasLandLease('0xabcdefabcdefabcdefabcdefabcdefabcdefabcd', ['-100,100'])
+
+      expect(mockLeaseFetch.fetch).toHaveBeenCalledTimes(2)
+    })
+  })
+
+  describe('getLeaseHoldersForParcels', () => {
+    describe('and no parcels are passed', () => {
+      it('should return an empty array without querying authorizations', async () => {
+        const result = await lands.getLeaseHoldersForParcels([])
+        expect(result).toEqual([])
+        expect(mockLeaseFetch.fetch).not.toHaveBeenCalled()
+      })
+    })
+
+    describe('and an authorization overlaps the requested parcels', () => {
+      beforeEach(() => {
+        mockLeaseFetch.fetch.mockResolvedValue({
+          ok: true,
+          json: () =>
+            Promise.resolve([
+              {
+                name: 'lease-1',
+                desc: '',
+                contactInfo: { name: 'tenant' },
+                addresses: ['0xLEASE1', '0xLease2'],
+                plots: ['10,20']
+              },
+              {
+                name: 'unrelated',
+                desc: '',
+                contactInfo: { name: 'other' },
+                addresses: ['0xunrelated'],
+                plots: ['99,99']
+              }
+            ])
+        })
+      })
+
+      it('should return only the lowercased addresses for overlapping plots', async () => {
+        const result = await lands.getLeaseHoldersForParcels(['10,20', '11,20'])
+        expect(new Set(result)).toEqual(new Set(['0xlease1', '0xlease2']))
+      })
+    })
+
+    describe('and the upstream fetch throws', () => {
+      beforeEach(() => {
+        mockLeaseFetch.fetch.mockRejectedValue(new Error('lease service unavailable'))
+      })
+
+      it('should swallow the error and return an empty array', async () => {
+        const result = await lands.getLeaseHoldersForParcels(['10,20'])
+        expect(result).toEqual([])
+      })
+    })
+  })
+})
